pkg/system: move maxTime init() back to Chtimes code
This code was moved to a separate file in fe5b34ba8828dc2f2f7db180a102cee360fec6e0,
but it's unclear why it was moved (as this file is not excluded on Windows).

Moving the code back into the chtimes file, to move it closer to where it's used.
